moved line_reader and its mocks to its own module
