+++ src/game/mod.rs
@@ -7,20 +7,7 @@ use entry::Entry;
 use stack::kind::{HasId, IsEmpty, Kind, KindId};
 use stack::Stack;
 use std::collections::HashMap;
-use std::io;
-pub trait LineReader: Default + Clone {
-    fn read_line(&self, input: &mut String);
-}
-
-
-
-#[derive(Debug, Default, Clone)]
-pub struct STDInReader {}
-impl LineReader for STDInReader {
-    fn read_line(&self, input: &mut String) {
-        io::stdin().read_line(input).unwrap();
-    }
-}
+use crate::line_reader::LineReader;
 
 pub struct Game<TR: LineReader> {
     stacks: Vec<Stack>,
@@ -213,8 +200,8 @@ impl<TLR: LineReader + Default + Clone> Game<TLR> {
 
 #[cfg(test)]
 mod tests {
-    use std::cell::RefCell;
     use super::*;
+    use crate::line_reader::{LineReader, MockLineReader, STDInReader};
 
     #[test]
     fn test_get_stages() {
@@ -227,36 +214,4 @@ mod tests {
     }
 
 
-    #[test]
-    fn test_mock_line_reader() {
-        #[derive(Clone)]
-        struct MockLineReader {
-            index: RefCell<usize>,
-            lines: Vec<String>
-        }
-
-
-        impl Default for MockLineReader {
-            fn default() -> Self {
-                MockLineReader { index: RefCell::new(0) , lines: vec!["1 2".to_string(), "3 4".to_string()]}
-            }
-        }
-        impl LineReader for MockLineReader {
-            fn read_line(&self, input: &mut String) {
-                let mut index = self.index.borrow_mut();
-                let line = self.lines.get(*index);
-                *index = *index + 1;
-                *input = line.unwrap().clone();
-            }
-        }
-
-        let mock = MockLineReader::default();
-        let mut input = String::new();
-        mock.read_line(&mut input);
-        assert!(input.eq("1 2"));
-        mock.read_line(&mut input);
-        assert!(input.eq("3 4"));
-
-
-    }
 }
