+++ src/lib.rs
@@ -1,1 +1,2 @@
 pub mod game;
+pub mod line_reader;
