+++ src/main.rs
@@ -1,5 +1,5 @@
 use sorting_game::game::Game;
-use sorting_game::game::STDInReader;
+use sorting_game::line_reader::STDInReader;
 fn main() {
     Game::<STDInReader>::play();
 }
